avatars with cache

+++ src/import/helpers/import-avatar.ts
@@ -1,41 +1,57 @@
 import { spawn } from "node:child_process";
-import { SOURCE_AVATARS_URL } from "../config";
+import { CACHE_FOLDER } from "../config";
 import { unlink } from "node:fs/promises";
 
 export const importAvatar = async (
-  avatar: string,
-  path: string,
+  url: string,
+  name: string,
   inputType: string,
   outputType: string
 ) => {
-  const inputPath = `${path}.${inputType}`;
-  const outputPath = `${path}.${outputType}`;
+  const path = `./data/${name}`;
+  const cacheKey = `avatar-${name}`;
+  const cache = Bun.file(`${CACHE_FOLDER}/${cacheKey}.json`);
 
-  const image = await fetch(`${SOURCE_AVATARS_URL}${avatar}`);
-  const blob = await image.blob();
-  await Bun.write(inputPath, blob);
-  const inputSize = Bun.file(inputPath).size;
+  if (cache.size === 0) {
+    const inputPath = `${path}.${inputType}`;
+    const outputPath = `${path}.${outputType}`;
 
-  const proc = spawn("ffmpeg", ["-i", inputPath, outputPath], {
-    shell: true,
-    env: { ...process.env },
-  });
+    const image = await fetch(url);
+    const blob = await image.blob();
+    await Bun.write(inputPath, blob);
+    const inputSize = Bun.file(inputPath).size;
 
-  await new Promise((resolve, reject) => {
-    proc.on("exit", (code) => {
-      if (code === 0) return resolve(true);
-      reject(code);
+    const proc = spawn("ffmpeg", ["-i", inputPath, outputPath], {
+      shell: true,
     });
-  });
 
-  await unlink(inputPath);
+    await new Promise((resolve, reject) => {
+      proc.on("exit", (code) => {
+        if (code === 0) return resolve(true);
+        reject(code);
+      });
+    });
 
-  const webpFile = Bun.file(outputPath);
-  const outputSize = webpFile.size;
+    await unlink(inputPath);
 
-  const base64 = Buffer.from(await webpFile.arrayBuffer()).toString("base64");
+    const webpFile = Bun.file(outputPath);
+    const outputSize = webpFile.size;
 
-  await unlink(outputPath);
+    const base64 = Buffer.from(await webpFile.arrayBuffer()).toString("base64");
 
-  return { base64, inputSize, outputSize };
+    await unlink(outputPath);
+
+    await Bun.write(
+      `${CACHE_FOLDER}/${cacheKey}.json`,
+      JSON.stringify({
+        base64,
+        inputSize,
+        outputSize,
+      })
+    );
+
+    return { base64, inputSize, outputSize };
+  }
+
+  return await cache.json();
 };
